Polish code structure
