+++ cert_pdf/vars.py
@@ -1,0 +1,4 @@
+import uuid
+
+TOKEN = str(uuid.uuid4())
+JOB_LOG = dict()
